Tests bot_get functions in test_bot_get.py
